+++ EveUtils.Client/Views/Killmails/KillmailAttackerLineView.axaml.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace EveUtils.Client.Views.Killmails;
+
+public partial class KillmailAttackerLineView : UserControl
+{
+    public KillmailAttackerLineView() => InitializeComponent();
+}