+++ EveUtils.Client/Views/Killmails/KillmailItemLineView.axaml.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace EveUtils.Client.Views.Killmails;
+
+public partial class KillmailItemLineView : UserControl
+{
+    public KillmailItemLineView() => InitializeComponent();
+}