+++ EveUtils.Client/Views/Killmails/KillmailLinkedRunView.axaml.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace EveUtils.Client.Views.Killmails;
+
+public partial class KillmailLinkedRunView : UserControl
+{
+    public KillmailLinkedRunView() => InitializeComponent();
+}